# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gratification API",
        "description": "Documentação da API do Gratification",
        "contact": {
            "email": "suporte@gratification.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Listar todas as empresas",
                "operationId": "22631014ece23858d19a7ef1c89f0e78",
                "responses": {
                    "200": {
                        "description": "Lista de todas as empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "docs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nome": {
                                                        "type": "string",
                                                        "example": "Empresa X"
                                                    },
                                                    "razao_social": {
                                                        "type": "string",
                                                        "example": "Razão Social X"
                                                    },
                                                    "status": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Criar uma nova empresa",
                "operationId": "edd395f6abdaa55cb0284b3b198d7ad6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "logo": {
                                        "type": "string",
                                        "example": "logo.png"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Empresa X"
                                    },
                                    "razao_social": {
                                        "type": "string",
                                        "example": "Razão Social X"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "empresa@x.com"
                                    },
                                    "telcom": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "telcel": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "12.345.678/0001-90"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empresa criada com sucesso!"
                                        },
                                        "docs": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Empresa X"
                                                },
                                                "razao_social": {
                                                    "type": "string",
                                                    "example": "Razão Social X"
                                                },
                                                "status": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Exibir uma empresa específica",
                "operationId": "1ab6ccd465b58367f4064aa417a103be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes de uma empresa específica",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "docs": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Empresa X"
                                                },
                                                "razao_social": {
                                                    "type": "string",
                                                    "example": "Razão Social X"
                                                },
                                                "status": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Atualizar uma empresa",
                "operationId": "b61661cfb086a49afd6ebebf15ed6011",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "logo": {
                                        "type": "string",
                                        "example": "logo_updated.png"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Empresa Y"
                                    },
                                    "razao_social": {
                                        "type": "string",
                                        "example": "Razão Social Y"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "empresa@y.com"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empresa atualizada com sucesso!"
                                        },
                                        "docs": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Empresa X"
                                                },
                                                "razao_social": {
                                                    "type": "string",
                                                    "example": "Razão Social X"
                                                },
                                                "status": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Excluir uma empresa",
                "operationId": "56b2aac859ceb472597b45436c7f5f12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empresa excluída com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}